requirements.txt
openai>=1.30.0
faiss-cpu>=1.7.4
pyodbc>=5.1
psycopg2-binary>=2.9
SQLAlchemy>=2.0
pandas>=2.1
pydantic>=2.5
PyYAML>=6.0
Jinja2>=3.1
python-dotenv>=1.0
tenacity>=8.2
numpy>=1.26
pyarrow>=15.0
sentence-transformers>=2.6
pytest>=8.0